lib/dl_core_testing/pyproject.toml

[tool.poetry]
name = "datalens-core-testing"
version = "0.0.1"
description = ""
authors = ["DataLens Team <datalens-opensource@yandex-team.ru>"]
packages = [{include = "dl_core_testing"}]
license = "Apache 2.0"
readme = "README.md"


[tool.poetry.dependencies]
Flask = ">=2.0.3"
Werkzeug = ">=2.0.3"
aiohttp = ">=3.8.1"
attrs = ">=22.2.0"
marshmallow = ">=3.19.0"
psycopg2-binary = ">=2.9.6"
pytest = ">=7.2.2"
python = ">=3.10, <3.12"
requests = ">=2.28.2"
shortuuid = ">=1.0.11"
sqlalchemy = ">=1.4.46, <2.0"
uwsgi = ">=2.0.21"
statcommons = {path = "../statcommons"}
datalens-api-commons = {path = "../dl_api_commons"}
datalens-utils = {path = "../dl_utils"}
datalens-constants = {path = "../dl_constants"}
datalens-testing = {path = "../dl_testing"}
datalens-configs = {path = "../dl_configs"}
datalens-core = {path = "../dl_core"}
datalens-app-tools = {path = "../dl_app_tools"}
datalens-db-testing = {path = "../dl_db_testing"}

[tool.poetry.plugins."dl_core.connectors"]
testing = "dl_core_testing.connector:TestingCoreConnector"

[tool.poetry.group.tests.dependencies]
pytest = ">=7.2.2"
docker = ">=6.0.1"  # Here is to prevent installation in prod deps as transitive dependency

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = []



[datalens_ci]
skip_test = true

[tool.mypy]
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
